+++ projects/ngx-uploader-directive/CHANGELOG.md
@@ -8,9 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 - Fixed error event emittion.
+- Fixed events emittion on invalid request id.
 
 ### Changed
 - Readme.
+- Changed error response to HttpErrorResponse
+- On rejection Http Error response will be returned
 
 ## [1.1.3] - 2020-05-03